+++ protocols/substreams/ethereum-ekubo-v3/src/lib.rs
@@ -1,5 +1,6 @@
 mod abi;
 mod addresses;
 mod details_store;
+mod genesis;
 mod modules;
 mod pb;